requirements.txt
# Twilio SDK – SMS, Voice, VoIP integration
twilio

# Stripe – Issuing virtual cards + Google Wallet push provisioning
stripe

# JWT verification for the dojo_bridge Control Plane API
# NestJS issues HS256 JWTs; Odoo verifies them here (never contacts Firebase directly)
PyJWT>=2.8.0

# Data Analysis for HRMS Dashboard
pandas
numpy

# Reporting & Barcode Engine Fixes
# We use the [bracket] syntax to ensure the graphics extras are installed
reportlab[renderPM,pycairo]
rlPyCairo

# QR Code generation for marketing cards
qrcode[pil]

# Standard Odoo dependencies (if not already handled by the platform) 
Pillow
babel